make ProtectedRoute component reusable
